requirements.txt
Django>=5.0,<6
celery>=5.3,<6
redis>=5
psycopg2-binary>=2.9
gunicorn>=22
pillow
whitenoise>=6.6
python-dateutil>=2.9
